updated models in profile + sports_event app to get float instead of integer

+++ sports_event_project/profiles/models.py
@@ -12,7 +12,7 @@ class Profile(models.Model):
     city = models.CharField(max_length=150)
     state = models.CharField(max_length=100)
     zipcode = models.CharField(max_length=100)
-    lat = models.IntegerField(blank=True, null=True)
-    lng = models.IntegerField(blank=True, null=True)
+    lat = models.FloatField(blank=True, null=True)
+    lng = models.FloatField(blank=True, null=True)
 
 # Create your models here.
